Change notes referencing relevant tests to use the `data-test` ReSpec feature.
The test-manifest HTML generation can extract these from the spec, and update the HTML with back-references to the relevant part of the spec.

For example, as is done for [YAML-LD tests](https://json-ld.github.io/yaml-ld/tests/manifest.html). Note References entry in most test cases that links back to the normative text.

Although a fair amount of work, this can be done with other specs and tests, and becomes valuable for the Implementation Report. Ideally, each test would be referenced by one or more locations in the spec, and each normative statement would reference one or more tests.